added proper file structure and used best practices for react
